WIP new versioning
Add notes about the meaning of Info.plist keys - I always forget this stuff.

Remove code for setting CFBundleBuildVersion - This is not a real
Info.plist key — a Google of it only gives the MacDown repo and the
script that it references.

TODO why is this reverting to some weird value the second time I run
locally after a clean?